backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
sqlalchemy
aiosqlite
python-jose[cryptography]
google-auth
requests
httpx